+++ ts/sdk/src/trade/update.rs
@@ -1,8 +1,12 @@
-use std::collections::HashMap;
+use std::collections::{hash_map::Entry, HashMap};
 
 use bs58_fixed_wasm::Bs58Array;
 use inf1_core::inf1_ctl_core::{
-    accounts::pool_state::PoolState,
+    accounts::{
+        lst_state_list::LstStatePackedList,
+        packed_list::PackedList,
+        pool_state::{PoolState, PoolStatePacked},
+    },
     keys::{LST_STATE_LIST_ID, POOL_STATE_ID},
     typedefs::lst_state::LstState,
 };
@@ -14,11 +18,15 @@ use crate::{
     interface::{Account, AccountMap, B58PK},
     missing_acc_err,
     pda::controller::create_raw_pool_reserves_ata,
+    sol_val_calc::Calc,
     trade::{Pair, PkPair},
-    utils::{balance_from_token_acc_data, try_find_lst_state},
+    utils::{balance_from_token_acc_data, token_supply_from_mint_data, try_find_lst_state},
     Inf, Reserves,
 };
 
+/// Returned accounts are deduped
+///
+/// @throws
 #[wasm_bindgen(js_name = accountsToUpdateForTrade)]
 pub fn accounts_to_update_for_trade(
     inf: &Inf,
@@ -192,4 +200,57 @@ fn update_reserves(
             .ok_or_else(|| acc_deser_err(&reserves_addr))?,
     });
     Ok(())
+}
+
+impl Inf {
+    pub(crate) fn update_ctl_accounts(
+        &mut self,
+        fetched: &HashMap<B58PK, Account>,
+    ) -> Result<(), JsError> {
+        let [pool_state_acc, lst_state_list_acc] = [POOL_STATE_ID, LST_STATE_LIST_ID].map(|pk| {
+            fetched
+                .get(&B58PK::new(pk))
+                .ok_or_else(|| missing_acc_err(&pk))
+        });
+        let pool_state_acc = pool_state_acc?;
+        let lst_state_list_acc = lst_state_list_acc?;
+
+        let pool = PoolStatePacked::of_acc_data(&pool_state_acc.data)
+            .ok_or_else(|| acc_deser_err(&POOL_STATE_ID))?;
+        let PackedList(lst_state_list) = LstStatePackedList::of_acc_data(&lst_state_list_acc.data)
+            .ok_or_else(|| acc_deser_err(&LST_STATE_LIST_ID))?;
+        lst_state_list.iter().for_each(|s| {
+            let s = s.into_lst_state();
+            // Initialize sol value calc and indiv LST data if newly added LST
+            if let Entry::Vacant(entry) = self.lsts.entry(s.mint) {
+                // TODO: we are ignoring Calc::new() error here
+                // so that we dont brick our stuff from adding a new unsupported SPL LST.
+                // We maybe want to handle this error properly instead
+                if let Ok(calc) = Calc::new(&s, &self.spl_lsts) {
+                    entry.insert((calc, None));
+                }
+            }
+        });
+        // TODO: maybe cleanup removed LSTs from self.lsts?
+
+        self.pool = pool.into_pool_state();
+        self.lst_state_list_data = lst_state_list_acc.data.as_slice().into();
+
+        Ok(())
+    }
+
+    pub(crate) fn update_lp_token_supply(
+        &mut self,
+        fetched: &HashMap<B58PK, Account>,
+    ) -> Result<(), JsError> {
+        let lp_mint_acc = fetched
+            .get(&B58PK::new(self.pool.lp_token_mint))
+            .ok_or_else(|| missing_acc_err(&self.pool.lp_token_mint))?;
+        let lp_token_supply = token_supply_from_mint_data(&lp_mint_acc.data)
+            .ok_or_else(|| acc_deser_err(&self.pool.lp_token_mint))?;
+
+        self.lp_token_supply = Some(lp_token_supply);
+
+        Ok(())
+    }
 }
